requirements.txt
flask
flask-sqlalchemy
flask-limiter
python-dotenv
PyPDF2
spacy
wordfreq
pyspellchecker
google-genai
gunicorn
